requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
matplotlib          # For charts (optional)
plotly              # For interactive visual charts (optional)
reportlab           # For PDF generation (optional)
